requirements/base.txt
Django==1.6.2
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
django-rest-swagger==0.1.14
djangorestframework==2.3.13
logutils==0.3.3
plist==0.2
South==0.8.2
